added transaction repository

+++ stock-management-backend/src/main/java/com/ofss/repository/TransactionRepository.java
@@ -9,11 +9,11 @@ import java.util.List;
 @Repository
 public interface TransactionRepository extends JpaRepository<Transaction,Integer> {
     // Get all transactions for a customer
-    List<Transaction> findByCustomerId(Long customerId);
+    List<Transaction> findByCustomerId(int customerId);
 
     // Get all transactions for a stock
-    List<Transaction> findByStockId(Long stockId);
+    List<Transaction> findByStockId(int stockId);
 
     // Get transactions for customer + stock
-    List<Transaction> findByCustomerIdAndStockId(Long customerId, Long stockId);
+    List<Transaction> findByCustomerIdAndStockId(int customerId,int stockId);
 }
